refactor(dashboard): extract row activation policy
